pyproject.toml
[project]
name = "wilma-mcp"
version = "0.1.0"
description = "MCP server for Wilma (Finnish school communication platform)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=0.4.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
wilma-mcp = "wilma_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/wilma_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
